+++ src/main/java/com/mapofmemory/global/common/CommonResponse.java
@@ -1,4 +1,4 @@
-package com.mapofmemory.global.dto;
+package com.mapofmemory.global.common;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
